chatbot_langchain_v1/pyproject.toml
[project]
name = "chatbot-langchain-v1"
version = "0.1.0"
description = "LangChain v1 Chatbot with Azure OpenAI + Pinecone"
authors = [
    { name = "Arvind Mali", email = "arvindmali5768@gmail.com" }
]

# Prevent Poetry from treating this like a pip package
package-mode = false

# Python version that works safely with all your dependencies
requires-python = ">=3.11,<3.13"

dependencies = [
    "python-dotenv>=1.0.1",
    "langchain>=1.0.0",
    "langchain-openai>=0.1.1",
    "langchain-community>=0.1.0",
    "langchain-text-splitters>=0.1.0",
    "pinecone-client>=3.0.0",
    "pypdf>=4.0.0",
    "streamlit>=1.30.0",
    "tiktoken>=0.7.0",
    "requests>=2.31.0",
    "azure-storage-blob (>=12.27.1,<13.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
